Apaired segments building for symmetric routing (step 1).
* Change: In Hurricane::BaseFlags, store flags in uint64_t instead of
    unsigned int because we start to need more than 32 different flags
    in some tools.
* New: In ::getString() & ::getRecord() templates, add support for
    std::array<>.
* Change: In CRL::ToolEngine, add support for timer (time & memory
    measurements) displaced from Katabatic. This way all ToolEngine
    can use this feature. The _postCreate() method display the
    memory just after ToolEngine allocation.
* Change: In Etesian::EtesianEngine, make use of the ToolEngine
    builtin timer (remove the local one). Forgot to call the base
    class _postCreate() and _preDestroy().
* Change: In Anabatic::AnabaticEngine, make use of the ToolEngine
    builtin timer (remove the local one).
* New: In Anabatic, new AutoSegments_Connecteds() collection. This
    Collection allows a deterministic walkthough *all* the AutoSegments
    connected either to source or target of one AutoSegment.
* New: In Anabatic::AutoContactTerminal::isEndPoint() to check if an
    AutoContactTerminal is the *only one* anchored on a RoutingPad,
    thus being a true "end point" and not a kind of feed-through.
* New: In Katana::KatanaEngine, added support for symmetric nets.
    Created new class DataSymmetric to store symmetric information
    of a net (mainly the paired AutoSegments).
      Added KatanaEngine::runSymmetricRouter(), for now only build
    the DataSymmetric informations. More to come...
* Change: In Katana::GraphicKatanaEngine::_runTest(), now perform
    symmetric information building the non-symmetric routing.

+++ anabatic/src/Constants.cpp
@@ -22,67 +22,67 @@ namespace Anabatic {
 
   using std::string;
 
-  const unsigned int  Flags::NoFlags             =  0;
+  const uint64_t  Flags::NoFlags             =  0;
 // Flags used for both objects states & functions arguments.
-  const unsigned int  Flags::Horizontal          = (1 <<  0);
-  const unsigned int  Flags::Vertical            = (1 <<  1);
-  const unsigned int  Flags::Source              = (1 <<  2);
-  const unsigned int  Flags::Target              = (1 <<  3);
-  const unsigned int  Flags::Invalidated         = (1 <<  4);
+  const uint64_t  Flags::Horizontal          = (1 <<  0);
+  const uint64_t  Flags::Vertical            = (1 <<  1);
+  const uint64_t  Flags::Source              = (1 <<  2);
+  const uint64_t  Flags::Target              = (1 <<  3);
+  const uint64_t  Flags::Invalidated         = (1 <<  4);
 // Flags for GCell objects states only.                        
-  const unsigned int  Flags::DeviceGCell         = (1 <<  5);
-  const unsigned int  Flags::HChannelGCell       = (1 <<  6);
-  const unsigned int  Flags::VChannelGCell       = (1 <<  7);
-  const unsigned int  Flags::StrutGCell          = (1 <<  8);
-  const unsigned int  Flags::MatrixGCell         = (1 <<  9);
-  const unsigned int  Flags::IoPadGCell          = (1 << 10);
-  const unsigned int  Flags::Saturated           = (1 << 11);
+  const uint64_t  Flags::DeviceGCell         = (1 <<  5);
+  const uint64_t  Flags::HChannelGCell       = (1 <<  6);
+  const uint64_t  Flags::VChannelGCell       = (1 <<  7);
+  const uint64_t  Flags::StrutGCell          = (1 <<  8);
+  const uint64_t  Flags::MatrixGCell         = (1 <<  9);
+  const uint64_t  Flags::IoPadGCell          = (1 << 10);
+  const uint64_t  Flags::Saturated           = (1 << 11);
 // Flags for Anabatic objects states only.                      
-  const unsigned int  Flags::DemoMode            = (1 <<  5);
-  const unsigned int  Flags::WarnOnGCellOverload = (1 <<  6);
-  const unsigned int  Flags::DestroyGCell        = (1 <<  7);
-  const unsigned int  Flags::DestroyBaseContact  = (1 <<  8);
-  const unsigned int  Flags::DestroyBaseSegment  = (1 <<  9);
+  const uint64_t  Flags::DemoMode            = (1 <<  5);
+  const uint64_t  Flags::WarnOnGCellOverload = (1 <<  6);
+  const uint64_t  Flags::DestroyGCell        = (1 <<  7);
+  const uint64_t  Flags::DestroyBaseContact  = (1 <<  8);
+  const uint64_t  Flags::DestroyBaseSegment  = (1 <<  9);
 // Flags for NetDatas objects states only.                      
-  const unsigned int  Flags::GlobalRouted        = (1 <<  5);
+  const uint64_t  Flags::GlobalRouted        = (1 <<  5);
 // Masks.                                      
-  const unsigned int  Flags::WestSide            = Horizontal|Target;
-  const unsigned int  Flags::EastSide            = Horizontal|Source;
-  const unsigned int  Flags::SouthSide           = Vertical  |Target;
-  const unsigned int  Flags::NorthSide           = Vertical  |Source;
-  const unsigned int  Flags::AllSides            = WestSide|EastSide|SouthSide|NorthSide ;
-  const unsigned int  Flags::EndsMask            = Source|Target;
-  const unsigned int  Flags::DirectionMask       = Horizontal|Vertical;
-  const unsigned int  Flags::DestroyMask         = DestroyGCell|DestroyBaseContact|DestroyBaseSegment;
-  const unsigned int  Flags::GCellTypeMask       = DeviceGCell|HChannelGCell|VChannelGCell|StrutGCell|MatrixGCell|IoPadGCell;
+  const uint64_t  Flags::WestSide            = Horizontal|Target;
+  const uint64_t  Flags::EastSide            = Horizontal|Source;
+  const uint64_t  Flags::SouthSide           = Vertical  |Target;
+  const uint64_t  Flags::NorthSide           = Vertical  |Source;
+  const uint64_t  Flags::AllSides            = WestSide|EastSide|SouthSide|NorthSide ;
+  const uint64_t  Flags::EndsMask            = Source|Target;
+  const uint64_t  Flags::DirectionMask       = Horizontal|Vertical;
+  const uint64_t  Flags::DestroyMask         = DestroyGCell|DestroyBaseContact|DestroyBaseSegment;
+  const uint64_t  Flags::GCellTypeMask       = DeviceGCell|HChannelGCell|VChannelGCell|StrutGCell|MatrixGCell|IoPadGCell;
 // Flags for functions arguments only.           
-  const unsigned int  Flags::Create              = (1 <<  5);
-  const unsigned int  Flags::WithPerpands        = (1 <<  6);
-  const unsigned int  Flags::WithSelf            = (1 <<  7);
-  const unsigned int  Flags::AboveLayer          = (1 <<  8);
-  const unsigned int  Flags::BelowLayer          = (1 <<  9);
-  const unsigned int  Flags::OpenSession         = (1 << 10);
-  const unsigned int  Flags::Realignate          = (1 << 11);
-  const unsigned int  Flags::NativeConstraints   = (1 << 12);
-  const unsigned int  Flags::ForceMove           = (1 << 13);
-  const unsigned int  Flags::WarnOnError         = (1 << 14);
-  const unsigned int  Flags::Topology            = (1 << 15);
-  const unsigned int  Flags::GlobalSegment       = (1 << 16);
-  const unsigned int  Flags::AllowTerminal       = (1 << 17);
-  const unsigned int  Flags::AllowLocal          = (1 << 18);
-  const unsigned int  Flags::IgnoreContacts      = (1 << 19);
-  const unsigned int  Flags::Propagate           = (1 << 20);
-  const unsigned int  Flags::Superior            = (1 << 21);
-  const unsigned int  Flags::DoglegOnLeft        = (1 << 22);
-  const unsigned int  Flags::DoglegOnRight       = (1 << 23);
-  const unsigned int  Flags::WithNeighbors       = (1 << 24);
-  const unsigned int  Flags::NoCheckLayer        = (1 << 25);
-  const unsigned int  Flags::HalfSlacken         = (1 << 26);
-  const unsigned int  Flags::NoGCellShrink       = (1 << 27);
-  const unsigned int  Flags::CParanoid           = (1 << 28);
-  const unsigned int  Flags::CheckLowDensity     = (1 << 29);
-  const unsigned int  Flags::CheckLowUpDensity   = (1 << 30);
-  const unsigned int  Flags::NoUpdate            = (1 << 31);
+  const uint64_t  Flags::Create              = (1 <<  5);
+  const uint64_t  Flags::WithPerpands        = (1 <<  6);
+  const uint64_t  Flags::WithSelf            = (1 <<  7);
+  const uint64_t  Flags::AboveLayer          = (1 <<  8);
+  const uint64_t  Flags::BelowLayer          = (1 <<  9);
+  const uint64_t  Flags::OpenSession         = (1 << 10);
+  const uint64_t  Flags::Realignate          = (1 << 11);
+  const uint64_t  Flags::NativeConstraints   = (1 << 12);
+  const uint64_t  Flags::ForceMove           = (1 << 13);
+  const uint64_t  Flags::WarnOnError         = (1 << 14);
+  const uint64_t  Flags::Topology            = (1 << 15);
+  const uint64_t  Flags::GlobalSegment       = (1 << 16);
+  const uint64_t  Flags::AllowTerminal       = (1 << 17);
+  const uint64_t  Flags::AllowLocal          = (1 << 18);
+  const uint64_t  Flags::IgnoreContacts      = (1 << 19);
+  const uint64_t  Flags::Propagate           = (1 << 20);
+  const uint64_t  Flags::Superior            = (1 << 21);
+  const uint64_t  Flags::DoglegOnLeft        = (1 << 22);
+  const uint64_t  Flags::DoglegOnRight       = (1 << 23);
+  const uint64_t  Flags::WithNeighbors       = (1 << 24);
+  const uint64_t  Flags::NoCheckLayer        = (1 << 25);
+  const uint64_t  Flags::HalfSlacken         = (1 << 26);
+  const uint64_t  Flags::NoGCellShrink       = (1 << 27);
+  const uint64_t  Flags::CParanoid           = (1 << 28);
+  const uint64_t  Flags::CheckLowDensity     = (1 << 29);
+  const uint64_t  Flags::CheckLowUpDensity   = (1 << 30);
+  const uint64_t  Flags::NoUpdate            = (1 << 31);
 
 
   Flags::~Flags ()
